Remove configmaps watcher
The injector will exit with a fatal error if the watcher can't reach the
apiserver momentarily. This is not necessarily desirable behaviour when
you are relying on the injector being available to inject sidecars into
pods. Not to mention the potential impact on apiserver requests while
the webhook server is down.

Rather than rework the watcher to tolerate errors, I thought, seen as
this is our fork and we don't use the feature (we mount the configmap and rely on kustomize
generators instead), we could just remove the watcher altogether.