Upgrade to admission.k8s.io/v1 (#6)
The v1beta1 version is deprecated and will be removed in 1.22+.

The v1 review validation checks the kind and apiVersion of the response,
which was absent, so I've added that in. I've also added a test which
checks those fields as well as the UID of the response which wasn't captured by
the other mutation tests.